requirements.txt
streamlit
openai
langchain-openai
langchain-core
faiss-cpu
PyMuPDF
python-dotenv
pathlib
numpy
